Make `Optional` types truly Optional
The codegen currently correctly emits optional data types as
`Optional[T]`. That's good, but what was missing was that these should
also have a default value of `None`
[[docs](https://docs.pydantic.dev/latest/concepts/fields/)].

This change now makes it such that optional fields are truly optional
and have a default value of `None` so that we can always upgrade schemas
safely.