Allow the codegen to emit the `null` type (#37)
Why
===

Some types have an explicit `Type.Null()` (as opposed to a
`Type.Optional`). That's allowed by the spec, so we should support it.

What changed
============

This change now correctly codegens `Type.Null()` (or in JSON-schema
parlance, `"type": "null"`) into Python's `None` type.

Test plan
=========

Amended the python file for this in the poetry venv, generating the
latest and greatest pid2 schema succeeded.