Also close the WebSocket when we close the session
When we're closing sessions, we do not close the WebSocket and instead
kinda rely on the server kicking us out due to inactivity. This makes
it racy for cases like connection retries (which we have disabled and
instead close any previous sessions), where we can have two concurrent
connections for the same session id. An invariant violation!

This change now also makes sure to close the WebSocket when the session
is closed.