Improve logging / exception hygiene (#57)
Why
===

We currently have some places where exceptions are not chained using the
`raise ... from` statement, but rather just the message is added to
another exception. That's suboptimal because we lose a lot of context.

What changed
============

This change now tries to use `raise ... from` in more places. Also, all
error logs that embed an exception into the log entry now use
`logging.exception` (which is equivalent plus it tacks the exception
into the `exc_info` field), and other logs that embed the exception
message into the log use `exc_info` instead (either setting it to `True`
if it is logging in an `except` block, or manually setting the
  `exc_info` to the `BaseException` instance). Also, it surfaces
  exception chains in the stringification for uncaught errors.

Test plan
=========

:eyes: on :dog:.