chore/schema handshake metadata pt2 (#74)
Why
===

Follow-up to #72 to account for `schema.json` files without metadata.

What changed
============

If the `handshakeSchema` field is defined, then the parameter is
required. Otherwise, the parameter is `Literal[None]`, which matches the
previous behavior of the default.

Due to the metadata field now being required, I had to remove the `=
None` default parameter. I think this is alright.

It'll mean that #73 will likely need to change to
`handshake_metadata_factory: HandshakeType | Callable[[],
Awaitable[HandshakeType]]` to avoid `async def stub() -> None: return
None` just to satisfy the async requirement. It's somewhat challenging
to capture the exact semantics we want here, but I think that's alright.

Test plan
=========

Do typechecks pass?