Add RiverServiceException (#76)
Why
===

Sometimes River has an issue, sometimes the service running River has an
issue. We should split those out so that people know which is which.
 
What changed
============

Added `RiverServiceException` to differentiate the two.
`RiverServiceException` also includes extra fields to include the
service and procedure.

Test plan
=========

CI/CD, this change is backwards compatible due to RiverServiceException
being a subclass of RiverException.