use empty context in restore budget task (#122)
Why
===

By default, new asyncio tasks hold onto a reference to the current
context. If the user of river has a lot of data in the context while the
restore budget task is created, that context cannot be GC'd until the
task exits. This can lead to accidental memory leaks.

As a followup we will replace this rate limiter with one that does not
require background tasks.

What changed
============

- Create a fresh context for restore budget task

Test plan
=========

- Shouldn't have any behavior impact here, we don't rely on the context
for budget restoration.
- We should see fewer memory leaks internally