Switch from context manager to a straight binding to avoid deadlock (#143)
Why
===

Attempting to run `uv run pytest -k test_ignore_flood_subscription`
hangs for a significant amount of time.

This is because in `__aexit__` we await on the server's `wait_closed()`,
which never completes because we still haven't called `server.close()`:

```
    async def __aexit__(...) -> None:
        self.ws_server.close()
        await self.ws_server.wait_closed()
```

An alternate approach would be to wrap the entire innards of the async
context in a `try: ... finally: server.close()`.

What changed
============

Reflowed `clientserver` text fixture to avoid a deadlock.

Test plan
=========

CI