Stricter error types (#141)
Why
===

Have the generated code return error types that actually satisfy the
`ErrorType`. Previously the union type for error types generated can be
values of `RiverUnknownValue` which is not an error type. This
introduces a `RiverUnknownError` and uses that in the error unions
instead. Also type adapter types are now more strictly typed.

# Testing

When switching the client code over, we should get errors about
`RiverUnknownError` not being handled and force you to fix them.

---------

Co-authored-by: Devon Stewart <devon.stewart@repl.it>